+++ CHANGELOG.md
@@ -29,6 +29,10 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
+## [4.4] `Hscroll` component
+
+- Added `Hscroll` component to layout utils
+
 ## [4.3] Add double spacing helpers
 
 - Add `--double` variants for `margin` helper classes